requirements.txt
google-genai>=1.5.0
gspread==6.1.2
google-auth==2.29.0
pandas>=2.2.0
python-dotenv>=1.0.0
Pillow>=10.3.0
streamlit>=1.30.0
beautifulsoup4>=4.12.0
requests>=2.31.0
yt-dlp>=2024.1.0
python-docx>=1.1.0
pdfplumber>=0.10.0
